modified the frontend for ApplicationView and application.properties
modified the frontend ApplicationView, modified the application.properties and added the CandidateApplicationService to backend. This is the improved version compared to the previous CandidateApplicationService, but is not yet successfully connected to frontend. I configured the database connection in application.properties to use localhost:5430, aligning it with the PostgreSQL service running in docker-compose.yaml.